+++ lib/new_relic/agent/instrumentation/active_record.rb
@@ -60,7 +60,14 @@ module NewRelic
             database = @config && @config[:database]
           end
 
-          segment = NewRelic::Agent::Transaction.start_datastore_segment(product, operation, collection, host, port_path_or_id, database)
+          segment = NewRelic::Agent::Transaction.start_datastore_segment(
+            product: product,
+            operation: operation,
+            collection: collection,
+            host: host,
+            port_path_or_id: port_path_or_id,
+            database_name: database
+          )
           segment._notice_sql(sql, @config, EXPLAINER)
 
           begin